fix: 恢复文件分组
