--- packages/amis-ui/src/components/Timeline/type.ts
@@ -1,9 +1,0 @@
-export type DotSize = 'sm' | 'md' | 'lg' | 'xl';
-
-export enum DirectionMode {
-  left = 'left',
-  right = 'right',
-  top = 'top',
-  bottom = 'bottom',
-  alternate = 'alternate'
-}
